refactor: move pages to screens folder
