+++ frontend/src/App.js
@@ -5,10 +5,10 @@ import { SWRConfig } from 'swr'
 import { getPositions } from './lib/utils'
 import { SOURCES_API, TRANSLATION_SOURCES_API, DB_VERSION_API } from './lib/consts'
 import { TranslationSourcesContext } from './lib/contexts'
-import Home from './components/Home'
-import SourceView from './components/SourceView'
-import LineView from './components/LineView'
-import LineRedirect from './components/LineRedirect'
+import Home from './screens/Home'
+import SourceView from './screens/SourceView'
+import LineView from './screens/LineView'
+import LineRedirect from './helpers/LineRedirect'
 
 import './App.css'
