+++ frontend/src/screens/LineView.css
