+++ frontend/src/screens/LineView.js
@@ -10,13 +10,13 @@ import useSWR from 'swr'
 import { PAGE_API } from '../lib/consts'
 import { getDictionaryLink, getIssueUrl } from '../lib/utils'
 
-import Error from './Error'
-import Loader from './Loader'
-import LinkButton from './LinkButton'
-import IconButton from './IconButton'
-import TranslationBlock from './TranslationBlock'
-import Menu from './Menu'
-import MenuItem from './MenuItem'
+import Error from '../components/Error'
+import Loader from '../components/Loader'
+import LinkButton from '../components/LinkButton'
+import IconButton from '../components/IconButton'
+import TranslationBlock from '../components/TranslationBlock'
+import Menu from '../components/Menu'
+import MenuItem from '../components/MenuItem'
 
 import './LineView.css'
